requirements.txt
Django>=3.0,<4.0
psycopg2>=2.8
djangorestframework>=3.14.0